# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Frostbyte ETL Pipeline API
  version: 1.0.0
  description: Data-sovereign document ingestion and vector query API.
paths:
  /documents:
    post:
      summary: Upload a document
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '202':
          description: Accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentMetadata'
        '400':
          $ref: '#/components/responses/BadRequest'
  /documents/{id}:
    get:
      summary: Get document metadata
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentMetadata'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/chunks:
    get:
      summary: List all chunks for a document
      operationId: listChunks
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chunk'
        '404':
          $ref: '#/components/responses/NotFound'
  /collections:
    post:
      summary: Create a new collection
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: {type: string}
                  created_at: {type: string, format: date-time}
        '409':
          $ref: '#/components/responses/Conflict'
  /collections/{name}/query:
    post:
      summary: Query a collection
      operationId: queryCollection
      parameters:
        - name: name
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/pipeline/status:
    get:
      summary: Get pipeline status
      operationId: getPipelineStatus
      tags: [pipeline]
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Pipeline status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PipelineStatus'}
        '401':
          description: Unauthorized

  /api/v1/config:
    patch:
      summary: Update pipeline configuration
      operationId: patchConfig
      tags: [pipeline]
      security: [{bearerAuth: []}]
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConfigPatchRequest'}
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PipelineStatus'}
        '400':
          description: Validation error

  /api/v1/tenants:
    get:
      summary: List tenants
      operationId: listTenants
      tags: [tenants]
      security: [{bearerAuth: []}]
      parameters:
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20, maximum: 100}}
      responses:
        '200':
          description: Tenant list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantsList'}

  /api/v1/tenants/{id}:
    get:
      summary: Get tenant by ID
      operationId: getTenant
      tags: [tenants]
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Tenant detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tenant'}
        '404':
          description: Not found

  /api/v1/documents:
    get:
      summary: List documents
      operationId: listDocuments
      tags: [documents]
      security: [{bearerAuth: []}]
      parameters:
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20, maximum: 100}}
        - {name: tenantId, in: query, schema: {type: string}}
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DocumentsList'}

  /api/v1/documents/{id}/chain:
    get:
      summary: Get document chain of custody
      operationId: getDocumentChain
      tags: [documents]
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: Chain of custody
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChainResponse'}
        '404':
          description: Not found

  /api/v1/verification/test:
    post:
      summary: Run verification test suite
      operationId: runVerificationTest
      tags: [verification]
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [testType]
              properties:
                testType:
                  type: string
                  enum: [red-team, compliance, penetration]
      responses:
        '200':
          description: Verification results
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VerificationResult'}
        '400':
          description: Invalid test type

  /health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, enum: [healthy]}
                  timestamp: {type: string, format: date-time}
components:
  schemas:
    DocumentUploadRequest:
      type: object
      required: [filename, content]
      properties:
        filename: {type: string}
        content:  {type: string, format: binary}
        metadata: {type: object, additionalProperties: true}
    DocumentMetadata:
      type: object
      properties:
        id:         {type: string, format: uuid}
        filename:   {type: string}
        status:     {type: string, enum: [pending, processing, completed, failed]}
        created_at: {type: string, format: date-time}
    Chunk:
      type: object
      properties:
        chunk_id:    {type: string, format: uuid}
        document_id: {type: string, format: uuid}
        content:     {type: string}
        embedding:   {type: array, items: {type: number}}
    CollectionCreateRequest:
      type: object
      required: [name]
      properties:
        name:        {type: string, pattern: '^[a-z0-9_-]{3,32}$'}
        description: {type: string}
    QueryRequest:
      type: object
      required: [vector, top_k]
      properties:
        vector: {type: array, items: {type: number}}
        top_k:  {type: integer, minimum: 1, maximum: 100}
    QueryResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              chunk_id:    {type: string}
              score:       {type: number}
              document_id: {type: string}
              content:     {type: string}
    PipelineStatus:
      type: object
      properties:
        mode: {type: string, enum: [online, offline, dual]}
        batch_size: {type: integer}
        model: {type: string}
        throughput: {type: number, format: float}
        nodes:
          type: array
          items:
            type: object
            properties:
              id: {type: string}
              status: {type: string}
              metrics:
                type: object
                properties:
                  rate: {type: number, nullable: true}
                  latency: {type: number, nullable: true}
    ConfigPatchRequest:
      type: object
      properties:
        mode: {type: string, enum: [online, offline, dual]}
        batch_size: {type: integer, minimum: 1, maximum: 256}
        model: {type: string}
    TenantsList:
      type: object
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Tenant'}
        total: {type: integer}
    Tenant:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        state: {type: string}
        config: {type: object}
        config_version: {type: integer}
        created_at: {type: string, format: date-time}
    DocumentsList:
      type: object
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/DocumentListItem'}
        total: {type: integer}
    DocumentListItem:
      type: object
      properties:
        id: {type: string, format: uuid}
        tenant_id: {type: string}
        filename: {type: string}
        status: {type: string}
        modality: {type: string}
        metadata: {type: object}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    ChainResponse:
      type: object
      properties:
        document_id: {type: string, format: uuid}
        events:
          type: array
          items:
            type: object
            properties:
              action: {type: string}
              timestamp: {type: string, format: date-time}
              performed_by: {type: string, nullable: true}
              metadata: {type: object}
    VerificationResult:
      type: object
      properties:
        score: {type: integer}
        details:
          type: array
          items:
            type: object
            properties:
              test: {type: string}
              passed: {type: boolean}
              message: {type: string}
    Error:
      type: object
      required: [error_code, message]
      properties:
        error_code: {type: string}
        message:    {type: string}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'